requirements.txt
Django==2.0.6
Pillow==5.1.0
certifi==2018.4.16
chardet==3.0.4
click==6.7
cloudinary==1.12.0
decorator==4.3.0
dj-database-url==0.5.0
django-registration-redux==2.4
django-widget-tweaks==1.4.2
future==0.16.0
geocoder==1.38.1
idna==2.7
mock==2.0.0
pbr==4.0.4
python-decouple==3.1
pytz==2018.4
ratelim==0.1.6
requests==2.19.1
six==1.11.0
urllib3==1.23
